requirements.txt
streamlit
pyyaml